refactor: change code location
createSnsUser, validateSnsUserDuplicate 메서드를 SnsAPIService에서 UserService로 옮김
