+++ src/main/java/com/example/backend/service/user/SnsAPIService.java
@@ -1,11 +1,8 @@
 package com.example.backend.service.user;
 
 //import com.example.backend.controller.TestKakaoToken;
-import com.example.backend.domain.User;
 import com.example.backend.domain.enumType.UserType;
-import com.example.backend.dto.login.SnsJoinRequestDTO;
 import com.example.backend.dto.user.SnsUserDTO;
-import org.springframework.transaction.annotation.Transactional;
 
 //import java.io.IOException;
 
@@ -15,9 +12,6 @@ public interface SnsAPIService {
     Long getSnsId(UserType userType, String accessToken);
 
     SnsUserDTO getUserInfo(UserType userType, String accessToken);
-
-    @Transactional
-    User createSnsUser(UserType userType, SnsJoinRequestDTO joinDTO);
 
     void unlink(UserType userType, String accessToken);
