+++ src/main/java/com/example/backend/service/user/UserService.java
@@ -1,12 +1,16 @@
 package com.example.backend.service.user;
 
 import com.example.backend.domain.User;
+import com.example.backend.domain.enumType.UserType;
+import com.example.backend.dto.login.SnsJoinRequestDTO;
 import com.example.backend.dto.user.UserDefaultJoinRequestDTO;
 import com.example.backend.dto.user.UserModifyRequestDTO;
 
 public interface UserService {
 
     User createDefaultUser(UserDefaultJoinRequestDTO userDTO); //회원가입
+
+    User createSnsUser(UserType userType, SnsJoinRequestDTO joinDTO);
 
     void changeToWithdrawnUser(User user); //회원 탈퇴
